feat: add chip selection functionality to TimeSeriesView component
